# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phantom Business SMS Gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OtpSendInput {
  /** Phone number in E.164 format (e.g. +12025551234) */
  phone: string;
  /**
     * Optional app identifier for multi-app gateways
     * @nullable
     */
  appId?: string | null;
}

export interface OtpSendResult {
  success: boolean;
  requestId: string;
  expiresAt: string;
  /**
     * The send channel used (modem, email-sms, dev)
     * @nullable
     */
  channel?: string | null;
}

export interface OtpVerifyInput {
  phone: string;
  /** The 6-digit OTP code */
  code: string;
  /** @nullable */
  requestId?: string | null;
}

export interface OtpVerifyResult {
  verified: boolean;
  attemptsRemaining: number;
  /**
     * Reason for failure if not verified
     * @nullable
     */
  reason?: string | null;
}

export type OtpLogStatus = typeof OtpLogStatus[keyof typeof OtpLogStatus];


export const OtpLogStatus = {
  pending: 'pending',
  sent: 'sent',
  delivered: 'delivered',
  failed: 'failed',
  verified: 'verified',
  expired: 'expired',
} as const;

export interface OtpLog {
  id: number;
  requestId?: string;
  phone: string;
  status: OtpLogStatus;
  /** @nullable */
  channel?: string | null;
  /** @nullable */
  appId?: string | null;
  attemptsUsed: number;
  createdAt: string;
  expiresAt: string;
  /** @nullable */
  verifiedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface GatewayStatus {
  online: boolean;
  /** Active send channel (modem, email-sms, dev) */
  channel: string;
  modemConnected?: boolean;
  /** @nullable */
  modemPort?: string | null;
  smtpReady?: boolean;
  lastChecked: string;
  /**
     * GSM signal strength 0-31 if modem connected
     * @nullable
     */
  signal?: number | null;
}

export type GatewayConfigChannel = typeof GatewayConfigChannel[keyof typeof GatewayConfigChannel];


export const GatewayConfigChannel = {
  modem: 'modem',
  'email-sms': 'email-sms',
  dev: 'dev',
} as const;

export interface GatewayConfig {
  id: number;
  channel: GatewayConfigChannel;
  otpLength: number;
  otpExpirySeconds: number;
  maxAttempts: number;
  rateWindowSeconds: number;
  maxPerWindow: number;
  /** @nullable */
  modemPort?: string | null;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
  /**
     * Carrier email-to-SMS gateway domain (e.g. txt.att.net, tmomail.net, vtext.com)
     * @nullable
     */
  smsGatewayDomain?: string | null;
  /** @nullable */
  senderName?: string | null;
  updatedAt?: string;
}

export type GatewayConfigInputChannel = typeof GatewayConfigInputChannel[keyof typeof GatewayConfigInputChannel];


export const GatewayConfigInputChannel = {
  modem: 'modem',
  'email-sms': 'email-sms',
  dev: 'dev',
} as const;

export interface GatewayConfigInput {
  channel?: GatewayConfigInputChannel;
  otpLength?: number;
  otpExpirySeconds?: number;
  maxAttempts?: number;
  rateWindowSeconds?: number;
  maxPerWindow?: number;
  /** @nullable */
  modemPort?: string | null;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpPassword?: string | null;
  /**
     * Carrier email-to-SMS gateway domain (e.g. txt.att.net, tmomail.net, vtext.com)
     * @nullable
     */
  smsGatewayDomain?: string | null;
  /** @nullable */
  senderName?: string | null;
}

export interface GatewayStats {
  totalSent: number;
  totalVerified: number;
  totalFailed: number;
  totalExpired: number;
  successRate: number;
  todaySent: number;
  todayVerified: number;
  recentLogs?: OtpLog[];
}

export interface ApiKey {
  id: number;
  /** API key value (only shown in full on creation) */
  key: string;
  name: string;
  appId: string;
  /**
     * SMS sender ID shown to recipients (max 11 chars for GSM). Overrides global config.
     * @nullable
     */
  senderName?: string | null;
  /**
     * URL to POST webhook events to
     * @nullable
     */
  webhookUrl?: string | null;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastUsedAt?: string | null;
}

export interface CreateApiKeyInput {
  name?: string;
  appId: string;
  /**
     * SMS sender ID for this app (max 11 chars)
     * @nullable
     */
  senderName?: string | null;
  /** @nullable */
  webhookUrl?: string | null;
}

export interface UpdateApiKeyInput {
  name?: string;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  webhookUrl?: string | null;
  isActive?: boolean;
}

export interface GatewayTestInput {
  phone: string;
}

export interface GatewayTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  channel?: string | null;
}

export type ListOtpLogsParams = {
limit?: number;
offset?: number;
status?: ListOtpLogsStatus;
};

export type ListOtpLogsStatus = typeof ListOtpLogsStatus[keyof typeof ListOtpLogsStatus];


export const ListOtpLogsStatus = {
  pending: 'pending',
  sent: 'sent',
  delivered: 'delivered',
  failed: 'failed',
  verified: 'verified',
  expired: 'expired',
} as const;

export type RevokeApiKey200 = {
  success: boolean;
};