# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phantom Business SMS Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: number;
  /** API key value (only shown in full on creation) */
  key: string;
  name: string;
  appId: string;
  /**
     * SMS sender ID shown to recipients (max 11 chars for GSM). Overrides global config.
     * @nullable
     */
  senderName?: string | null;
  /**
     * URL to POST webhook events to
     * @nullable
     */
  webhookUrl?: string | null;
  isActive: boolean;
  createdAt: Date;
  /** @nullable */
  lastUsedAt?: Date | null;
}