# lib/api-zod/src/generated/types/createApiKeyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phantom Business SMS Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApiKeyInput {
  name?: string;
  appId: string;
  /**
     * SMS sender ID for this app (max 11 chars)
     * @nullable
     */
  senderName?: string | null;
  /** @nullable */
  webhookUrl?: string | null;
}