# lib/api-zod/src/generated/types/updateApiKeyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phantom Business SMS Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateApiKeyInput {
  name?: string;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  webhookUrl?: string | null;
  isActive?: boolean;
}